% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{Add new geom to the ggmap to plot hurricane wind radii chart for a single storm observation to a map.
This function builds a custom geom for ggplot2 that can be used to add the hurricane wind radii chart for a single storm observation to a map}
\usage{
geom_hurricane(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{storm data specifying maximum wind speeds reached at specified wind intensities in knots.  These data should be in long format (one row for each wind speed intensity. The maximum wind speeds}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{x}{longitude of the location}

\item{y}{latitude of the location}

\item{r_ne}{max wind speed reached in the north east quadrant of the location}

\item{r_se}{max wind speed reached in the south east quadrant of the location}

\item{r_sw}{max wind speed reached in the south west quadrant of the location}

\item{r_nw}{max wind speed reached in the north west quadrant of the location}

\item{scale_radii}{wind radii give the maximum radial extent of winds of a certain direction in each quadrant. In some circumstances, you may want to plot a certain percentage of the full wind radii, to give a better idea of average exposures in each quadrant}
}
\value{
plots circular polygons for each of the three wind speed intensities based on the maximum wind speeds and in each quadrants.  scale_radii input will scale the polygon based on the input (1 will be the default value)

plots circular polygons for each of the three wind speed intensities based on the maximum wind speeds and in each quadrants.  scale_radii input will scale the polygon based on the input (1 will be the default value)
}
\description{
Add new geom to the ggmap to plot hurricane wind radii chart for a single storm observation to a map.
This function builds a custom geom for ggplot2 that can be used to add the hurricane wind radii chart for a single storm observation to a map
}
\examples{
map_plot <- ggmap::get_map("Lousiana", zoom = 6, maptype = "toner-background")
map_plot \%>\%
 ggmap(extent = "device") +
 geom_hurricane(data = storm_observation,
                aes(x = longitude, y = latitude,
                    r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
                    color = wind_speed, fill = wind_speed, scale_radii=0.8)) +
 scale_color_manual(name = "Wind speed (kts)",
                    values = c("red", "orange", "yellow")) +
 scale_fill_manual(name = "Wind speed (kts)",
                   values = c("red", "orange", "yellow"))
}
