% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\docType{data}
\name{GeomHurricane}
\alias{GeomHurricane}
\title{Create Geom to add the hurricane wind radii chart for a single storm observation to a map
This function builds a custom geom for ggplot2 that can be used to add the hurricane wind radii chart for a single storm observation to a map}
\format{An object of class \code{GeomHurricane} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomHurricane
}
\arguments{
\item{data}{storm data specifying maximum wind speeds reached at specified wind intensities in knots.  These data should be in long format (one row for each wind speed intensity. The maximum wind speeds in each of the four quadrants (ne, se, nw, sw) should be in separate columns for each of the wind speed intensities.}

\item{x}{longitude of the location}

\item{y}{latitude of the location}

\item{r_ne}{max wind speed reached in the north east quadrant of the location}

\item{r_se}{max wind speed reached in the south east quadrant of the location}

\item{r_sw}{max wind speed reached in the south west quadrant of the location}

\item{r_nw}{max wind speed reached in the north west quadrant of the location}

\item{scale_radii}{wind radii give the maximum radial extent of winds of a certain direction in each quadrant. In some circumstances, you may want to plot a certain percentage of the full wind radii, to give a better idea of average exposures in each quadrant}
}
\value{
plots circular polygons for each of the three wind speed intensities based on the maximum wind speeds and in each quadrants.  scale_radii input will scale the polygon based on the input (1 will be the default value)
}
\description{
Create Geom to add the hurricane wind radii chart for a single storm observation to a map
This function builds a custom geom for ggplot2 that can be used to add the hurricane wind radii chart for a single storm observation to a map
}
\keyword{datasets}
